+++ dietcalc/views/special_food/calc.py
@@ -11,7 +11,7 @@ def special_food_calc(request):
     if request.method == 'POST':
         form = SpecialFoodCalcForm(request.POST)
         special_food = Food.objects.get(id=request.POST['food'])
-        count_special_food = (100 * (int(request.POST['weight']) * float(request.POST['norm_per_kg'])))/float(special_food.protein)
+        count_special_food = (100 * (float(request.POST['weight']) * float(request.POST['norm_per_kg'])))/float(special_food.protein)
         return render(request, 'dietcalc/spec_food/calc_view.html', {'form': form, 'menuhome': MenuHome.objects.all(), 'count_special_food': count_special_food, 'special_food_name': special_food})
     else:
         form = SpecialFoodCalcForm()
